Add a Perfomance Monitoring doc for Go
This doc provides information on performance monitoring and a net/http
middleware example to get our gophers up and running with Airbrake
Performance Monitoring.
Updates performance pages to link to the go performance doc.
Updates the title for the Performance Monitoring for Rails doc, now that
we have more than one language to get started with.
Extracts the Performance Dashboard Features to it's own doc.